+++ transaction-user-server/src/main/java/com/dream/seata/user/server/service/tcc/UserAmountInfoForTccService.java
@@ -17,5 +17,5 @@ public interface UserAmountInfoForTccService {
      * @param deductionAmount 扣减金额
      * @return
      */
-    Result settlement(String userUid, Integer version, BigDecimal deductionAmount);
+    Result settlement(String userUid, Long version, BigDecimal deductionAmount);
 }
